Create wager.sol
wager.sol is a simple king-of-the-hill type betting smart contract, where players have the chance to win a pot of ether by outbidding other players. Each time a new bet is recorded, the remaining game time increases by one hour. The pot winner receives 90% of the prize, the contract creator receives 5%, and the remaining 5% is used as the starting pot of the next game.